Replace the explicit image name with a variable
The date of the raspbian buster images were wrong, as the downloaded image had a
date in june, and the convert script, had a date of march, thus the command did
not work out of the box.

This fix moves the script to use a variable for the images, to mitigate these
issues for the future.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>